requirements.txt
pynsq==0.6.7
simplejson==3.8.0